backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
alembic==1.13.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
email-validator==2.1.1
python-multipart==0.0.9
razorpay==1.4.2
qrcode[pil]
reportlab
slowapi
pandas==2.0.0
scikit-learn
joblib
numpy
ollama
textblob